third_party/blink/web_tests/fast/dom/shadow/alternate-stylesheets-expected.txt

CONSOLE WARNING: line 5: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Title attribute should be ignored for style elements in shadow trees.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(root.querySelector('div')).color is "rgb(0, 128, 0)"
PASS getComputedStyle(alternate).color is "rgb(0, 128, 0)"
PASS host.shadowRoot.styleSheets[0].title is null
PASS successfullyParsed is true

TEST COMPLETE

Should be green
